Changes Updated

+++ AdminPanel/adminPanel/src/main/java/com/nagarro/adminPanel/dao/RechargeWalletDao.java
@@ -2,6 +2,6 @@ package com.nagarro.adminPanel.dao;
 
 public interface RechargeWalletDao {
 
-	public void rechargeInDB(String customerID, int rechargeAmount);
+	public void rechargeInDB(String customerID, double rechargeAmount);
 
 }
