+++ AdminPanel/adminPanel/src/main/java/com/nagarro/adminPanel/services/RechargeWalletServices.java
@@ -2,6 +2,6 @@ package com.nagarro.adminPanel.services;
 
 public interface RechargeWalletServices {
 
-	public void rechargeWallet(String customerID, int rechargeAmount);
+	public void rechargeWallet(String customerID, double rechargeAmount);
 
 }
